requirements.txt
asgiref==3.2.10
autopep8==1.5.3
boto3==1.12.27
botocore==1.15.27
Django==3.1.2
django-debug-toolbar==2.2
django-extra-views==0.13.0
django-filter==2.3.0
django-mathfilters==1.0.0
django-random-queryset==0.1.3
django-storages==1.9.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docutils==0.15.2
jmespath==0.9.5
Pillow==10.0.0
psycopg2-binary==2.9.6
pycodestyle==2.6.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.8
whitenoise==5.0.1
xlwt==1.3.0
